# Fix performance (remove a lot of unnecessary allocs and drops)