> Ipld allocated and dropped with short living time

from #7

### Before:
```txt
Hello World Decode:
===================
libipld  : 243 ns

Realistic Decode Tests:
=======================
citm_catalog.json.dagcbor      libipld  : 9.55 ms (34.18 MB/s)
canada.json.dagcbor            libipld  : 11.01 ms (91.46 MB/s)
twitter.json.dagcbor           libipld  : 3.27 ms (117.36 MB/s)
```

### After:
```txt
Hello World Decode:
===================
libipld  : 263 ns

Realistic Decode Tests:
=======================
citm_catalog.json.dagcbor      libipld  : 4.61 ms (70.81 MB/s)
canada.json.dagcbor            libipld  : 6.69 ms (150.66 MB/s)
twitter.json.dagcbor           libipld  : 2.53 ms (151.70 MB/s)
```

- citm_catalog.json.dagcbor **x2**
- canada.json.dagcbor **x1.64**
- twitter.json.dagcbor **x1.29**